RDC-V1/requirements.txt
psutil>=5.9.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
requests>=2.31.0